requirements.txt
# Flask-related dependencies
Flask==2.2.2
Flask-Cors==3.0.10

# PyTorch and torchvision
torch==2.0.1+cpu  # Adjust the version to your needs
torchvision==0.15.2+cpu
torchaudio==2.0.1+cpu
-f https://download.pytorch.org/whl/cpu

# Transformers library
transformers==4.28.1

# PIL (Python Imaging Library)
pillow==9.5.0
gunicorn

# phi-related dependencies
phi==0.1.4
phidata
# Machine learning and preprocessing
scikit-learn==1.2.2
albumentations==1.3.0
albumentations[imgaug]==1.3.0
albumentations[pytorch]==1.3.0

# PDF processing
pdfplumber==0.5.28

# Google Generative AI
google-generativeai==0.2.0

# Data manipulation
pandas==1.5.3
numpy==1.24.2

# NLP processing
spacy==3.5.0
